feat: 消息后台自毁

+++ src/components/ChatMessage/MessageRow.tsx
@@ -1,4 +1,4 @@
-import React, { RefObject, useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import React, { RefObject, useCallback, useEffect, useMemo, useRef } from 'react'
 import { RowProps } from './MessageVariableSizeList'
 import { useMessageStore } from '@/stores/message'
 import { useTooltipsStore } from '@/stores/tooltips'
@@ -17,9 +17,8 @@ interface MessageRowProps {
 const MessageRow: React.FC<RowProps & MessageRowProps> = ({ index, setItemSize, selectChange, onSelect, el }) => {
 	const msgStore = useMessageStore()
 	const tooltipStore = useTooltipsStore()
-	const msg = useMemo(() => msgStore.messages[index], [index])
+	const msg = useMemo(() => msgStore.messages[index], [msgStore.messages, index])
 	const itemRef = useRef<HTMLDivElement | null>(null)
-	const [read, setRead] = useState<boolean>(true)
 
 	// 查找回复消息
 	const replyMessage = useCallback((msg_id: number) => msgStore.messages.find((v) => v?.msg_id === msg_id), [])
@@ -48,56 +47,32 @@ const MessageRow: React.FC<RowProps & MessageRowProps> = ({ index, setItemSize, 
 	const handlerObserver = (entries: IntersectionObserverEntry[]) => {
 		entries.forEach(async (entry) => {
 			if (entry.isIntersecting) {
-				// !isMe(msg.sender_id) &&
-				msgStore.readMessage([msg])
-				setRead(true)
-				// 如果消息是阅后即焚
-				if (msg?.is_burn_after_reading === MessageBurnAfterRead.YES) {
-					const time = msgStore.userInfo?.preferences?.open_burn_after_reading_time_out ?? 10
-					// 添加进去
-					UserStore.add(UserStore.tables.read_destroy, {
-						uid: msg.uid,
-						msg_id: msg?.msg_id,
-						read_time: Date.now(),
-						self_destruct_time: time
-					})
-					const timer = setTimeout(() => {
-						msgStore.deleteMessage(msg.msg_id)
-						UserStore.delete(UserStore.tables.read_destroy, 'uid', msg.uid)
-						clearTimeout(timer)
-					}, time * 1000)
-				}
-
-				// 取消监听
-				itemRef.current && ob.unobserve(itemRef.current)
+				msgStore.updateReads(msg)
 			}
 		})
 	}
 	const ob = useMemo(() => new IntersectionObserver(handlerObserver, { root: el.current! }), [])
 
-	const is_read = useMemo(() => msg?.is_read === MESSAGE_READ.READ, [read])
+	const is_read = useMemo(() => msg?.is_read === MESSAGE_READ.READ, [msg])
 
 	useEffect(() => {
 		// 已读就不需要再多做处理, 下面是处理未读的消息 || isMe(msg.sender_id)
 		if (!is_read) {
-			setRead(false)
 			requestAnimationFrame(() => {
 				setTimeout(() => {
 					itemRef.current && ob.observe(itemRef.current)
 				}, 1000)
 			})
-		} else if (msg.is_burn_after_reading === MessageBurnAfterRead.YES) {
+		}
+		//  防止有漏掉的消息
+		if (is_read && msg.is_burn_after_reading === MessageBurnAfterRead.YES) {
 			msgStore.deleteMessage(msg.msg_id)
 			UserStore.delete(UserStore.tables.read_destroy, 'uid', msg.uid)
 		}
 	}, [])
 
 	return (
-		<div
-			ref={itemRef}
-			className={clsx('h-auto', !msg?.is_tips && !is_read && !read && 'bg-gray-200')}
-			style={{ zIndex: 1 }}
-		>
+		<div ref={itemRef} className={clsx('h-auto', !msg?.is_tips && !is_read && 'bg-gray-200')} style={{ zIndex: 1 }}>
 			<List noChevron mediaList className="my-0">
 				<ListItem
 					key={index}
